Refactor test structure and improve reusability
Consolidated shared test logic into `AbstractIT` for reusability and streamlined client creation in `Clients.kt`. Renamed test classes with `IT` suffix to follow integration test naming conventions. Updated dependencies to include `junit-jupiter-params`.
